Update patch set 2
Patch Set 2:

Because:

1. The comment in Identity shows the expected format of `` in `Identity.Name`:

2. The parsing code does exactly the same mapping:

3. The comment in Entity says `// indexed by Identity.Name`, but that is not true, it is indexed by ``  (Which is implicitly assumed to be the same).

So we have the documentation and the parsing code assume `Identity.Name = `, and the only place that assumption is not held is in `NewEntity`, where `Identity.Name = `. This was actually discovered in a bug we had in our code, and resurfaced again in a code-review last week. So this should be changed in the sake of consistency.

Patch-set: 2
Tag: mailMessageId=<CAN_AA3b3HPyrp-P4_9pNOpJgTMPtyje1DHj6o=PYPBUVi87_Qg@mail.gmail.com>