Update patch set 2
Patch Set 2:

The formatting and links did not survive my e-mail reply. Here it is again:
Because:

  1. The comment in Identity shows the expected format of `packet.UserId.Id` in `Identity.Name`:  https://github.com/golang/crypto/blob/e8f229864d71a49e5fdc4a9a134c5f85c4c33d64/openpgp/keys.go#L37
  2. The parsing code does exactly the same mapping:
https://github.com/golang/crypto/blob/e8f229864d71a49e5fdc4a9a134c5f85c4c33d64/openpgp/keys.go#L350
  3. The comment in Entity says `// indexed by Identity.Name`, but that is not true, it is indexed by `Identity.UserId.Id`  (Which is implicitly assumed to be the same).
 https://github.com/golang/crypto/blob/e8f229864d71a49e5fdc4a9a134c5f85c4c33d64/openpgp/keys.go#L29


So we have the documentation and the parsing code assume `Identity.Name = Identity.UserId.Id`, and the only place that assumption is not held is in `NewEntity`, where `Identity.Name = Identity.UserId.Name`.

This was actually discovered in a bug we had in our code, and resurfaced again in a code-review last week.

So this should be changed in the sake of consistency.

Patch-set: 2
Reviewer: Gerrit User 26463 <26463@62eb7196-b449-3ce5-99f1-c037f21e1705>