Update patch set 2
Patch Set 2:

I'm not clear what this change is attempting.  Ctrl-D is a terminal's typical `eof', see output of `stty -a', and so at the start of the line can be treated as EOF.  When not at the start, but before the end, then readline implementations often bind it to `delete-char' that deletes one character forward, see `bind -p' output.  These appear to be what's currently implemented.

Ctrl-C is a terminal's typical `intr' character, causing a SIGINT to be sent to a process.  That ascii(7) gives it a end-of-... name doesn't imply to me that SSH here should handle it as Ctrl-D.

Patch-set: 2
CC: Gerrit User 10961 <10961@62eb7196-b449-3ce5-99f1-c037f21e1705>