Update patch set 3
Patch Set 3:

> Patch Set 2: Code-Review+2
> 
> We do it wrong also for RSA (and signatures I believe). I considered fixing it in https://golang.org/cl/100844 but it seemed too much complexity to be worth it if the extra zeroes were allowed by the other implementations.
> 
> Mike, did you find an implementation that does strict parsing?

Filippo, the issue wasn't strict parsing, but rather the inputs to the GnuPG key ID, which include bit length in the calculation. This can result in a set of subkeys and signatures that refer to a nonexistent master key (or, more understandably, they use a key ID that doesn't match the actual key ID of the master key they belong to).

Copying a portion of a mail I sent to Adam:

===
For a weekend personal project, I'm using the openpgp package to generate ECC keys to import, ascii-armored, into GnuPG. I was generating a primary and subkey, self-signing both, and then exporting the whole Entity. When I imported into GPG, the signatures' key IDs didn't match the primary's key ID; moreover, when I tried calculating the primary key ID myself, I also got different results.

I figured out that this line [https://github.com/golang/crypto/blob/master/openpgp/packet/public_key.go#L247] is the offender. It's assuming the bit length of the version || X || Y sequence is always the number of bytes * 8. But according to RFC 6637, "[the] adjusted underlying field size is the underlying field size that is rounded up to the nearest 8-bit boundary," which I think the original implementer might have read as being the size of the three concatenated components, when it's actually the size of the X and Y coordinates' byte sequences. So the rounding happens for the X and the Y, but not for the constant version 4, which takes up an extra 3 bits. This was causing the serialized length [https://github.com/golang/crypto/blob/master/openpgp/packet/public_key.go#L421] to disagree with what GnuPG expects (e.g., serialized length of 520 bits for NIST P-256 versus the 515 explicitly mentioned in RFC 6637 -- (256 + 256 + 3) rounded = 520 versus 256 rounded + 256 rounded + 3 = 515). Since that total length is a component of the GPG key ID, the Go library is miscalculating the key fingerprint, and all the signatures and subkeys refer to a nonexistent primary key (among many other problems).
===

Let me know if you have more questions.

Patch-set: 3
Reviewer: Gerrit User 26282 <26282@62eb7196-b449-3ce5-99f1-c037f21e1705>