internal/chacha20: implement the cipher.Stream interface
SIMD implementations of ChaCha20 (such as CL 35842) interleave block
computations in order to achieve high performance. This means that
they produce more than 64 bytes of output at a time. Unfortunately
when encrypting small amounts of data (such as Poly1305 keys) the
current interface to ChaCha20 forces the additional encrypted blocks
of output to be discarded and recomputed later since it does not
maintain any state. This additional overhead slows down the encryption
of small amounts of data.

This CL makes the ChaCha20 implementation stateful, caching key and
counter values and buffering encrypted blocks. It also now implements
the high level cipher.Stream interface which makes it more consistent
with other ciphers in the standard library's crypto package.

For now this change just modifies the implementation, consumers of the
API will be modified in future CLs.

This CL also adds a benchmark to the package to make adding later
optimizations easier.

Change-Id: I28ca7947904e9d79debe2d5aac6623526fe5e595