Update patch set 3
Patch Set 3:

I tested it, the performance of go implementation now matches the assembly version. Should I nuke the assembly and refine the go implementation now that the performance matches? I would need to intrinsify the functions that perform the multiword arithmetic for the go implementation to work. I don't think the design of the go implementation in current form is right for example, I created another version for mulWW in poly1305 package although mulWW is already implemented in math/big because it was not exported. Can I export mulWW in math/big and also add other required functions to math/big if we decide to choose the go implementation?

Patch-set: 3
Reviewer: Gerrit User 25702 <25702@62eb7196-b449-3ce5-99f1-c037f21e1705>