Update patch set 7
Patch Set 7:

I've been working on a similar change for s390x (CL 35842). As part of that effort I want to improve the performance of the ChaCha20-Poly1305 AEAD for small payloads by allowing the ChaCha20 key stream to be buffered and I've been working on this in CL 104856. Once the generic ChaCha20-Poly1305 code is updated to use this new API it will mean we can generate a large block of key stream values (e.g. 256 bytes) in one go and then XOR it with the Poly1305 key and up to 192 bytes of data without generating more key stream values.

If CL 104856 is accepted it would mean you would only need the block4 code in your assembly (always generating 256 bytes at a time and then buffering unused bytes), so I think it will be worth waiting for that CL to go in and then reworking this code to fit into it. You should then see a performance gain at the 64 byte payload size.

Patch-set: 7
CC: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>