Update patch set 8
Patch Set 8: Run-TryBot+1

(5 comments)

> Not sure supporting a change of s.buf size is worth all those extra indexes in the complex slicing spots.

I agree, but it is hard to avoid unnecessary allocations if the struct provided for the assembly implementation is different to the generic one. For example, polyKey will escape in the AEAD which hurts performance on small messages quite a lot.

I'd like to leave it this way for now and maybe revisit once all the SIMD implementations are in.

Patch-set: 8
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1