Update patch set 9
Patch Set 9:

> This is great, thanks for working on this. Is there a plan to pulling it back into vendor for this release?

Yes, personally I'm hoping to re-vendor it once all the CLs to optimize ChaCha20 and Poly1305 that are currently under review are submitted. That might extend a little bit into the freeze, so we could re-vendor it on the 1st May or something like that too so we aren't picking up too many changes once the freeze starts.

Patch-set: 9