Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Hi Adam, adding you here as original author of "crypto/tls/generate_cert.go". Is there a particular reason why it was not done a part of public API?

This level of API is really quite high-level and not generally suitable for the standard library (which is where the generate_cert.go tool lives). Maybe it could be suitable for x/crypto, but it's not immediately obvious that such a simplification is going to be sufficiently useful for people that they won't end up using the crypto/x509 API directly.

Patch-set: 1
Reviewer: Gerrit User 5425 <5425@62eb7196-b449-3ce5-99f1-c037f21e1705>