Update patch set 1
Patch Set 1:

This would definitely need a proposal and a better package than "util", but I also agree that there isn't a clear role for this API. When generating one-off certificates one can hardcode most of the options offered here, making the code much smaller.

You might want to consider instead adding an example to x509.CreateCertificate that shows how to make a tls.Certificate (with for example a P-256 key specifically).

Patch-set: 1
CC: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>