Update patch set 3
Patch Set 3:

Based on the bug, someone misconfigured an OCSP server and we want to work around it? That's not a very compelling reason, by itself, for this change. However, I assume that everything else accepts it and Go is the odd one out enforcing this? If so, then there might be a case  that it's not a battle worth fighting.

Patch-set: 3
Reviewer: Gerrit User 5425 <5425@62eb7196-b449-3ce5-99f1-c037f21e1705>