Update patch set 6
Patch Set 6:

> Thanks for merging the patch, but could you explain why the OCSP server is misconfigured?  My reading of the spec suggests that what they're doing is okay, so I'd like to be set straight if that interpretation is wrong.

In this case, the responder is sending a chain to a root, but we don't want to validate to a root. If we did that then a different CA could sign OCSP responses for someone else. (Or, if we limited it to being the same root, you could still have different intermediate CAs signing each other's OCSP responses.)

Rather we need the trust to go to the issuer for the certificate that we're checking. So certificates above the issuer are superfluous and that responder is just wasting bandwidth.

One could have an OCSP responder that sends two or more certificates to reach the issuer. Technically I think that's within the specs, but I don't think it should be supported because I don't believe that the operational benefits to the CA are worth the complexity and I don't think I've ever seen a CA do it.

Lastly, there's also, somewhere, the idea that one can have a completely separate OCSP system with a different PKI and manually-configured clients to know about it. There one could reasonably want multiple certificates in an OCSP response, but we only try to implement most of the WebPKI and, in practice, people don't run web browsers like that. (And I'm not sure that anyone does it anywhere, expect maybe the Federal Bridge.)

So the PKIX RFCs err strongly on the side of complexity, but that results in an over-complex mess and must be subsetted. The trick is pick the right battles and, over time, try to bring the complexity down.

Patch-set: 6