poly1305: implement a subset of the hash.Hash interface
This CL adds the poly1305.Hash type which implements a
subset of hash.Hash interface. With Hash it is possible
to compute an authentication tag of data without copying
it into a single byte slice.

This commit modifies the reference/generic and the
AMD64 assembler but not the ARM implementation to support
an io.Writer interface.

Updates golang/go#25219

Change-Id: I7ee5a9eadd43387cf3cd887d734c625575eee47d