Update patch set 1
Patch Set 1:

(1 comment)

In CL 110297 we need to fall back on the generic implementation if the processor does not have the necessary features. Just something to consider.

Also, we should really be testing the generic implementation on all platforms (especially amd64) so that we know if we break it even if we are using an assembly implementation.

Patch-set: 1
Reviewer: Gerrit User 11990 <11990@62eb7196-b449-3ce5-99f1-c037f21e1705>