cryptotest: new package
cryptotest is a suite of tests for crypto/cipher implementations, like
nettest. It's meant to be used both by the stdlib and by community
implementations.

Also, fix a bug it found in chacha20 where it would reslice dst to the
length of src if it had enough capacity (due to the cap and len BCE hint).

Updates golang/go#25309

Change-Id: I9ee2383fbc24c5a68e217439d443095d7d317213