acme/autocert: support both RSA and ECDSA clients on the fly
GetCertificate has all the information it needs to know if a client
supports ECDSA in ClientHelloInfo. Deprecate and ignore ForceRSA, and
just obtain a RSA certificate on the fly when a client that doesn't
support ECDSA connects.

This changes the cache key format to have a "/rsa" suffix for RSA
certificates. The default (ForceRSA = false) cache key is unchanged,
so most DirCache instances will still be valid. Caches created with
ForceRSA set will be silently ignored and certificates reissued.

ECDSA support detection is more strict in following RFC 5246 than
crypto/tls, which ignores signature_algorithms.

DO NOT SUBMIT: missing a few tests

Change-Id: I70227747b563d6849cb693f83a950d57040b3f39