openpgp: read keys with revoked user ids
Drop the subloop for packets after the uid. Instead, use the upper loop
and only add the user id to the list of identities once we find a valid
self-signature.

The existing code was wrongly assuming that all UserID packets are
immediately followed by a Signature packet. According to RFC4880, this
is not the case.

See RFC4880 11.1:
> Immediately following each User ID packet, there are zero or more
> Signature packets.

This change will ensure that Entities that are not immediately followed
by a Signature packet are read without raising a StructuralError.
Instead, UserID packets that are not immediately followed by a self
signature will be ignored.

Better support for revoked user ids can be added in a follow-up patch.

Fixes golang/go#25850

Change-Id: Idde34b97429998f28e0c687171024e51ed959bf0