openpgp: read keys with revoked user ids
Drop the subloop for packets after the uid. Instead, use the upper loop
and only add the user id to the list of identities once we find a valid
self-signature.

The existing code was wrongly assuming that all UserID packets should be
immediately followed by a Signature packet. This is not the case for
revoked UserIDs.

Fixes golang/go#25850

Change-Id: Idde34b97429998f28e0c687171024e51ed959bf0