openpgp: read keys with revoked user ids
The existing code was wrongly assuming that:
 1. UserID packets must be immediately followed by a Signature packet
 2. Signature packets following a UserID are necessairly a self sig

1. See RFC4880 11.1:
> Immediately following each User ID packet, there are zero or more
> Signature packets.

2. The Signature packet following a UserID could be a revocation.

Drop the subloop for packets after the uid. Instead, use the upper loop
and only add the user id to the list of identities once we find a valid
self-signature.

This change will ensure that Entities that are not immediately followed
by a Signature packet are read without raising a StructuralError.
Instead, UserID packets that are not immediately followed by a self
signature will be ignored.

Maximum backwards compatibility is retained because revoked UserIDs are
not added to the Entity's identities.

In a follow-up patch, we could adopt one of the following solutions:
 - Solution A (subjectively cleaner API):
     + Add all UserIDs to the Entity's Identities, even if they are not
       followed by a self signature.
     + Add a Revocations slice to Identities
 - Solution B (maximum backwards compatibility):
     + Add a RevokedIdentities slice to to Entity
     + Add the revoked UserIDs to the Identity's RevokedIdentities

Note that Solution B does not handle cases where:
 - There is no self signature after the UserID packet
 - There is no revocations after the UserID packet

Fixes golang/go#25850

Change-Id: Idde34b97429998f28e0c687171024e51ed959bf0