openpgp: expose revoked user ids
As of Change-Id Idde34b97429998f28e0c687171024e51ed959bf0, entities with
revoked UserIDs no longer cause a StructuralError. Instead, they are
ignored.

In order to expose these ignored UserIDs in the API, two solutions are
considered:

- Solution A:
  + Add all UserIDs to the Entity's Identities, even if they are not
    followed by a self signature.
  + Add a Revocations slice to Identity.

- Solution B:
  + Add a RevokedIdentities slice to Entity.
  + Add a NeithierRevokedNorSelfSignedIdentities slice to Entity.
  + Add the revoked UserIDs to the Identity's RevokedIdentities.
  + Add the UserIDs that are neithier revoked nor self signed to
    the Identify

This patch is an implementation of solution A. It compromises on the
previous unstated/accidental guarantee that all Identities in an Entity
contain a valid Self-Signed signature.

Change-Id: I1d93358327b3bfd648131082354fbdb53a190758