Update patch set 1
Patch Set 1:

The docs should have the caveat "if the cipher.Block can be used concurrently". (Feel free to patch that in.) In practice most (all?) can on most (all?) architectures. It would be useful to map this out and document it. Filed https://github.com/golang/go/issues/25882

It's unfortunate that any slice passed into cipher.Block must escape. I assume it's because it's an interface and we can do nothing about it?

In this change you might want to use a sync.Pool, if you can show that it makes a difference in benchmarks. Do not introduce a hidden mutex, nobody would expect it.

Patch-set: 1
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>