Update patch set 23
Patch Set 23:

(1 comment)

Thanks for iterating on this change!

The spec says the signatures still have to be immediately after the User ID, so shouldn't we keep an inner loop and just move revocation code there?

Entity already has a Revocations slice, any reason not to put the revocations there? AFACT this is just about the order in which things are, so there is already a scenario in which we put a revoked Identity in Identities. (Whether that's wise, I'm not saying.)

Patch-set: 23
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>