openpgp: implement Cert Revocation Signature
The existing code would ignore user ids that are not immediately
followed by a self-signature. However, according to RFC4880 11.1,
they are valid:
> Immediately following each User ID packet, there are zero or more
> Signature packets.

This had the effect of also ignoring revoked UserIDs, depending on the ordering
of the packets. Consider the following scenarios:
 A: UID SELFSIG REV
 B: UID REV SELFSIG

In scenario A, the UID would be added to the Entity's Identities, but in
Scenario B, it wouldn't.

Bring consistenty by adding UIDs to the entity's identities whether or
not they are followed by a signature.

Implement Certification revocation signature as defined by
RFC4880 5.2.1.

Create a new Revocations slice in Identity and add Certification
revocation signatures in that slice.

Fixes golang/go#27602

Change-Id: Ia29914fbab5e49ae72b1851ea1b75e38245fec27