internal/subtle: add Google App Engine support
The new package subtle added in golang.org/cl/112236 breaks
compatibility with App Engine due to the import of unsafe.

This changes adds an App Engine alternative without using usafe.
The assumption is that overlapping slices always share the same base.

Tested with:

  $ go test -test.tags=appengine -v
  === RUN   TestAliasing
  --- PASS: TestAliasing (0.00s)
  PASS
  ok  	golang.org/x/crypto/internal/subtle	0.009s

I think it might be a terrible idea. Suggestions?

Change-Id: I2fc6b02a860b3ee11fa31652ba302fc7db9df153