Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > Patch Set 9:
> > 
> > According to https://tools.ietf.org/html/draft-ietf-acme-acme-12#section-7.1.3 the order status can have the value "ready", which I don't see mapped in the types.go file. Am I looking in the wrong place?
> 
> https://go-review.googlesource.com/c/crypto/+/86635/9/acme/acme.go references draft-09. The "ready" order status was added in draft-10 (current most draft is draft-11). I think you're looking in the right place and there has been drift between the draft RFC and the proposed implementation.
> 
> Adding support for the ready order status is a good idea. The Let's Encrypt ACME v2 API is going to be returning this state after this week's configuration update: https://community.letsencrypt.org/t/acmev2-order-ready-status/62866

I've also just noticed that section 7.1.4 (https://tools.ietf.org/html/draft-ietf-acme-acme-12#page-26) of the ACME spec (draft-12) also has an additional status that is not in this package: 'expired'.

An individual authorisation may be in an expired state as per:

---
status (required, string):  The status of this authorization.
      Possible values are: "pending", "valid", "invalid", "deactivated",
      "expired", and "revoked".
---

I've not actually seen this status used myself, but I guess if an order is not finalised within some reasonable amount of time, its authorisations may expire.

Patch-set: 9