Update patch set 7
Patch Set 7:

Hey folks, is there anything we can do to get this PR moving? We've got 5 more PRs lined up, one of which increases the (weak) default S2KCount but it depends on this, which I opened back in September.

If you're not maintaining crypto/openpgp that's cool but please let me know since there's at least one vibrant fork (ProtonMail) that we'd be happy to switch to.

Patch-set: 7