Update patch set 1
Patch Set 1: Run-TryBot+1 Code-Review+2

Thank you for this. I'm going to say that a test isn't needed because there's no externally-observable behaviour change: this is a typo in an internal function, but it didn't matter because that function was never used in a way where it mattered. Still, worth a small change to fix it.

Patch-set: 1
Reviewer: Gerrit User 5425 <5425@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2
Label: Run-TryBot=+1