Update patch set 7
Patch Set 7:

> alternately, you should wait on the goroutine that does ServeAgent to exit.

Good call. Just changed this up to use a WaitGroup. I also dropped the "invalid opcode" test case since that's actually a well-formed request (so it doesn't necessarily fit into this test code to begin with) and it could result in unexpected IO errors since it tries to write an agentFailure response back to the "client" that closes the connection before trying to read the response.

Ran the new version on my local system with a high count, so I've got some higher confidence that I've weeded out race conditions this time around.

Patch-set: 7