Update patch set 12
Patch Set 12:

I'm sorry for not getting to this review sooner, but I will be reverting the change because pkcs12 is a frozen package with intentionally limited functionality (its only purpose is to act as a compatibility layer for certificates encoded as legacy PKCS#12 files), and because this PR does not have a multi-certificate test.

Decode is effectively just a helper, and you can use ToPEM in place of DecodeAll (at the cost of a key marshal/unmarshal, which is not an expensive operation). I will add a note to the docs in that sense. I would also be happy to merge a PR adding an example for it, or adding a test with multiple certificate bags.

If you do want to have a DecodeAll helper, I suggest submitting a PR to https://github.com/SSLMate/go-pkcs12 which is less limited in scope and based on the same codebase.

Patch-set: 12
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>