Update patch set 20
Patch Set 20:

> Patch Set 20:
> 
> > Patch Set 20: Code-Review+2
> > 
> > LGTM, let's wait merging until release-branch.go1.12 forks.
> 
> Does that mean this will be in 1.12 or 1.13?

This is in x/crypto, so it's not tied directly to the releases, it's just another Go library. However, we keep a separate branch in case we need to apply patches to the version of this that we vendored into a release.

It will not be vendored into 1.12, but you can use it from 1.12 without issues.

Patch-set: 20
Reviewer: Gerrit User 11715 <11715@62eb7196-b449-3ce5-99f1-c037f21e1705>