Update patch set 4
Patch Set 4:

> 1. is there a reason this has to be in the SSH package?

I put in the SSH package as it requires internal structures, specifically there is no way to provide a custom mux/transport on a Conn struct and disable the initial banner exchange as an external package. 
I'd be open to moving it into it's own package and maintaining it on my GitHub since technically this is a vendor specific protocol and not defined in an RFC as part of SSH overall.
If you think it would make more sense how would you suggest this be implemented to export the required methods on Conn?

Patch-set: 4
CC: Gerrit User 31156 <31156@62eb7196-b449-3ce5-99f1-c037f21e1705>