Update patch set 9
Patch Set 9:

Is this still being actively worked on? If not, I'd be happy to pick up the review feedback/comments.

Regarding ACMEv1 vs v2, and which we should support as part of the crypto library, Let's Encrypt have recently announced that ACMEv1 is moving to EoL: https://community.letsencrypt.org/t/end-of-life-plan-for-acmev1/88430

Does this change whether we want to support the old protocol moving forward? As a maintainer of a widely used ACME client, we moved over to utilising the code in this CL a year ago now, and it has worked well. We have completely dropped support for ACMEv1.

Could we rename the old package to `acmev1` (or name the new one `acmev2`), and maintain both implementations for now if support for the old protocol is required?

Regarding whether the API surface for v1 vs v2 should be the same - I would vote 'no'. The protocols are sufficiently different and use different nouns to represent similar concepts that I think aligning the interfaces would cause more confusion than anything else.

From my understanding, this CL needs updating to use the new POST-as-GET behaviour, and resolving the v1 vs v2 issues. Is there anything else that people would like to see before this can be considered for merge?

Patch-set: 9