poly1305: rewrite the Go implementation with 64-bit limbs
The new code is meant to be readable without external references for
Poly1305, and explains the field logic.

[BENCHMARKS GENERIC]

The assembly is still... For the Go code to catch up with the assembly...

[BENCHMARKS GENERIC VS ASSEMBLY]

Hopefully this also avoids the need for an arm64 implementation.

[BENCHMARKS ARM64]

Since now the Go and the amd64 assembly use the same limb schedule,
drop the assembly initialize and finalize implementations, and make
the wrapper code match. It comes with a minor slowdown.

[BENCHMARKS AMD64]

Since poly1305/sum_generic.go was almost entirely rewritten, it's
probably best reviewed on gitiles.

This is the implementation published at
https://blog.filippo.io/a-literate-go-implementation-of-poly1305/

Change-Id: I74f9011d3ee317a43b05ae7f05d96081d08bffd3