Update patch set 3
Patch Set 3:

Armando asked me to post here because I've contributed to this codebase, a version of which is part of Zinc/WireGuard. I haven't read or reviewed this Go port of it yet to see how it compares, but in case of a question about copyright or something: here's my thumbs up on the Go CLA and whatever else you require along those lines.

Patch-set: 3
CC: Gerrit User 20556 <20556@62eb7196-b449-3ce5-99f1-c037f21e1705>